Fixed Installer

+++ install.sh
@@ -2,6 +2,11 @@ echo "Hunt Installer Script"
 echo '-------------------------------------------'
 echo
 
+if [ "$EUID" -ne 0]
+    then echo "Not running as root. Please use sudo and try again."
+    exit
+fi
+
 echo "Removing current installation, if any..."
 rm -rf /usr/local/src/Hunt
 rm -rf /usr/local/bin/hunt
@@ -15,7 +20,7 @@ echo "Done."
 sleep 1
 echo "Moving application to /usr/local/src/Hunt/..."
 mkdir /usr/local/src/Hunt
-cp src/* /usr/local/src/Hunt/
+cp -r src/* /usr/local/src/Hunt/
 echo "Done."
 
 sleep 1
